feat(api): more organized API
