+++ src/openai_responses/ext/__init__.py
@@ -1,0 +1,4 @@
+import httpx
+import respx
+
+__all__ = ["httpx", "respx"]
