+++ src/openai_responses/ext/httpx.py
@@ -1,0 +1,1 @@
+from httpx import *  # noqa: F403
