+++ src/openai_responses/ext/respx.py
@@ -1,0 +1,1 @@
+from respx import *  # noqa: F403
